Moved Balance caching from the `Account` to the `Book` and removed the account's `Split` list.
The balance caching was very slow, for two reasons.  Each account was being registered as an "interested party" for every transaction addition and removal from the book.  This meant that a huge multicast delegate was fired off for every transaction.  Also, the invalidating of an account's balance resulted in the notification of all parent accounts that it had changed, even if these parents accounts didn't have a chached total to invalidate.

Now, we are indexing each account's balance, and using a memoized total balance function.  The total balance set is invalidated along with the update of the balance.  This happens in a loop, rather than recursive event calls.

Notifying of added transactions will still happen, so that UI elements can repaint, new rows can be added, etc...  However, this will not likely be a huge multicast, since there should be just one or two UI elements that are registered at any given time.